unrool 3

+++ Guiao1/src/utils.c
@@ -6,5 +6,5 @@
 //não precisam de ser exatos, apenas que nos dê uma ordem de comparação da dsitância 
 //dos pontos.
 float euclidean_distance(float ponto1_x,float ponto1_y,float ponto2_x,float ponto2_y){
-    return  (float) (ponto2_x - ponto1_x)*(ponto2_x - ponto1_x)+ (ponto2_y - ponto1_y)*(ponto2_y - ponto1_y);
+    return  (ponto2_x - ponto1_x)*(ponto2_x - ponto1_x)+ (ponto2_y - ponto1_y)*(ponto2_y - ponto1_y);
 }
